requirements.txt
aiohttp==3.9.*
websockets==12.*
pyyaml==6.*
python-dotenv==1.0.*
web3==6.*
tenacity==8.*
python-telegram-bot==21.*
